+++ cmake/base_config.cmake
@@ -69,11 +69,10 @@ endif()
 # Enable/disable warnings
 if(WARNINGS)
   message(STATUS "Warnings enabled")
-  add_compile_options(-Wall -Wextra -Wunreachable-code -Wunused) # -Werror
+  add_compile_options(-Wall -Wextra -Wunreachable-code -Wunused -Werror)
 else()
+  message(STATUS "Warnings disabled")
   add_compile_options(-w)
-  set(XACC_CMAKE_CXX_FLAGS "\"-DBOOST_DISABLE_PRAGMA_MESSAGE=ON -w\"")
-  set(dashw_IF_NOT_WARNINGS "-w")
 endif()
 
 # Add linker paths to installed binary rpaths